Task4-5/requirements.txt
gradio
google-generativeai
python-dotenv
Pillow
numpy
scikit-learn
faiss-cpu
